# Faster traverse