--- a/Data/Sequence/Internal.hs
+++ b/Data/Sequence/Internal.hs
@@ -425,7 +425,9 @@ instance Foldable Seq where
 #endif
 
 instance Traversable Seq where
+#if __GLASGOW_HASKELL__
     {-# INLINABLE traverse #-}
+#endif
     traverse f' (Seq EmptyT) = pure (Seq EmptyT)
     traverse f' (Seq (Single (Elem x'))) =
         (\x'' -> Seq (Single (Elem x''))) <$> f' x'