--- a/Utils/Containers/Internal/Coercions.hs
+++ b/Utils/Containers/Internal/Coercions.hs
@@ -10,14 +10,22 @@ import Data.Coerce
 #endif
 
 infixl 8 .#
+infixr 9 #.
 #if __GLASGOW_HASKELL__ >= 708
 (.#) :: Coercible b a => (b -> c) -> (a -> b) -> a -> c
 (.#) f _ = coerce f
+
+(#.) :: Coercible c b => (b -> c) -> (a -> b) -> a -> c
+(#.) _ = coerce (\x -> x :: b) :: forall a b. Coercible b a => a -> b
 #else
 (.#) :: (b -> c) -> (a -> b) -> a -> c
 (.#) = (.)
+
+(#.) :: (b -> c) -> (a -> b) -> a -> c
+(#.) = (.)
 #endif
 {-# INLINE (.#) #-}
+{-# INLINE (#.) #-}
 
 infix 9 .^#