--- a/Utils/Containers/Internal/Coercions.hs
+++ b/Utils/Containers/Internal/Coercions.hs
@@ -10,22 +10,14 @@ import Data.Coerce
 #endif
 
 infixl 8 .#
-infixr 9 #.
 #if __GLASGOW_HASKELL__ >= 708
 (.#) :: Coercible b a => (b -> c) -> (a -> b) -> a -> c
 (.#) f _ = coerce f
-
-(#.) :: Coercible c b => (b -> c) -> (a -> b) -> a -> c
-(#.) _ = coerce
 #else
 (.#) :: (b -> c) -> (a -> b) -> a -> c
 (.#) = (.)
-
-(#.) :: (b -> c) -> (a -> b) -> a -> c
-(#.) = (.)
 #endif
 {-# INLINE (.#) #-}
-{-# INLINE (#.) #-}
 
 infix 9 .^#