So I applied the same techniques in the folds to the traverse, and I got similar speedups. On the multiplyup benchmark: 

```haskell
multiplyUp :: Seq.Seq Int -> Seq.Seq Int
multiplyUp = snd . flip runState 0 . traverse go where
  go x = do
    s <- State (\x -> (x + 1,x))
    return (s * x)
```

On a sequence of length 500000:

```
benchmarking 500000/multiplyUp/new
time                 137.6 ms   (129.7 ms .. 145.7 ms)
                     0.996 R²   (0.990 R² .. 1.000 R²)
mean                 138.8 ms   (136.4 ms .. 141.4 ms)
std dev              3.470 ms   (2.492 ms .. 4.484 ms)
variance introduced by outliers: 12% (moderately inflated)
             
benchmarking 500000/multiplyUp/old
time                 355.1 ms   (281.2 ms .. 407.7 ms)
                     0.995 R²   (0.983 R² .. 1.000 R²)
mean                 339.1 ms   (324.9 ms .. 346.0 ms)
std dev              12.30 ms   (0.0 s .. 12.63 ms)
variance introduced by outliers: 19% (moderately inflated)
```

And even old functions which use traverse internally get a speedup. `scanl (+) 0` for instance:

```
benchmarking 500000/scanl/new
time                 134.5 ms   (121.4 ms .. 156.0 ms)
                     0.979 R²   (0.945 R² .. 0.998 R²)
mean                 141.0 ms   (134.2 ms .. 148.6 ms)
std dev              9.440 ms   (6.655 ms .. 12.88 ms)
variance introduced by outliers: 13% (moderately inflated)
             
benchmarking 500000/scanl/old
time                 373.0 ms   (253.8 ms .. 471.4 ms)
                     0.988 R²   (0.957 R² .. 1.000 R²)
mean                 379.5 ms   (356.0 ms .. 396.6 ms)
std dev              25.81 ms   (0.0 s .. 29.52 ms)
variance introduced by outliers: 20% (moderately inflated)
```

The inlining of the inner functions also allows us to avoid coercion around `Elem`. The code is *much* longer, though.