@treeowl 2018-01-26
There seems to be some verbosity inherent in efficient traversal, thanks to the limitations of the type role system.