@treeowl 2018-01-27
Don't worry too much about it. traverse can be a weirdo. We should try to
cut down on it where we don't need it though.

On Jan 27, 2018 2:13 PM, "Donnacha Oisín Kidney" <notifications@github.com>
wrote:

> *@oisdk* commented on this pull request.
> ------------------------------
>
> In Data/Sequence/Internal.hs
> <https://github.com/haskell/containers/pull/516#discussion_r164279016>:
>
> > +        traverseTree
> +            :: Applicative f
> +            => (Node a -> f (Node b))
> +            -> FingerTree (Node a)
> +            -> f (FingerTree (Node b))
> +        traverseTree _ EmptyT = pure EmptyT
> +        traverseTree f (Single x) = Single <$> f x
> +        traverseTree f (Deep s pr m sf) =
> +            liftA3
> +                (Deep s)
> +                (traverseDigitN f pr)
> +                (traverseTree (traverseNodeN f) m)
> +                (traverseDigitN f sf)
> +        traverseDigitE
> +            :: Applicative f
> +            => (a -> f b) -> Digit (Elem a) -> f (Digit (Elem b))
>
> That slows it down again, but I may be able to figure out a FingerTree
> instance which keeps the speed isn't to do with Elem
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/516#discussion_r164279016>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzi_SORqU8oCo63W_4bPmfvE0CckVjdks5tO3VTgaJpZM4RtzAL>
> .
>