@oisdk 2018-01-27
Yeah, so this version delegates to an `Elem`-agnostic traverse on `FingerTree`, and it seems like the extra `fmap`s are elided, because it has the same performance on the benchmarks