@treeowl 2018-01-27
I'm not sure I quite buy it. What if that final `fmap` is expensive?