@treeowl 2018-01-27
I'm concerned about traversing all those `Elem`s, actually. Let's not do
that.

On Jan 27, 2018 2:25 PM, "Donnacha Oisín Kidney" <notifications@github.com>
wrote:

> Yeah, so this version delegates to an Elem-agnostic traverse on FingerTree,
> and it seems like the extra fmaps are elided, because it has the same
> performance on the benchmarks
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/516#issuecomment-361008621>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzi_dDJNk8f78FWUElA9GVK1H6fCQz8ks5tO3gygaJpZM4RtzAL>
> .
>