@oisdk 2018-01-27
Yeah, I agree. The extra `fmap`s were making me nervous (although I couldn't find a test to make them slower