@treeowl 2018-03-09
Hey, did you happen to check out whether `traverseWithIndex` needs any similar attention?