@oisdk 2018-03-10
I'm not sure—I *think* I did. I'll check later today, but I'd imagine it's fine: [this](https://github.com/haskell/containers/blob/fd1a7d351ba528dfcada745a1b8167642d599602/Data/Sequence/Internal.hs#L2907) specialisation is effectively doing the same as in `traverse`.